+++ .projen/tasks.json
@@ -26,7 +26,30 @@
     },
     "bundle": {
       "name": "bundle",
-      "description": "Prepare assets"
+      "description": "Prepare assets",
+      "steps": [
+        {
+          "spawn": "bundle:app/api.lambda"
+        }
+      ]
+    },
+    "bundle:app/api.lambda": {
+      "name": "bundle:app/api.lambda",
+      "description": "Create a JavaScript bundle from src/app/api.lambda.ts",
+      "steps": [
+        {
+          "exec": "esbuild --bundle src/app/api.lambda.ts --target=\"node14\" --platform=\"node\" --outfile=\"assets/app/api.lambda/index.js\" --sourcemap"
+        }
+      ]
+    },
+    "bundle:app/api.lambda:watch": {
+      "name": "bundle:app/api.lambda:watch",
+      "description": "Continuously update the JavaScript bundle from src/app/api.lambda.ts",
+      "steps": [
+        {
+          "exec": "esbuild --bundle src/app/api.lambda.ts --target=\"node14\" --platform=\"node\" --outfile=\"assets/app/api.lambda/index.js\" --sourcemap --watch"
+        }
+      ]
     },
     "clobber": {
       "name": "clobber",
