+++ src/app/api.lambda.ts
@@ -1,5 +1,5 @@
 import { default as serverlessExpress } from '@vendia/serverless-express';
-import { buildApp } from '../app';
+import { buildApp } from './app';
 
 const app = buildApp();
